New module `vie2`
For now, it contains one class `VIE2Solver`.
That class solves the Volterra integral equation of the second kind

    y_a(t) + \sum_b \int_0^t ds k_{ab}(t, s) y_b(s) = q_b(t)

w.r.t. y_a(t) on a uniform real time mesh using start-up and time-stepping
procedures based on the Gregory quadrature rule. The solution y_a(t) is,
in general, tensor-valued with an arbitrary shape and 'a' is the
corresponding multi-index.

The new module is accompanied by a unit test, which makes use of some
reference results derived in a Mathematica notebook `doc/vie2.nb`.