keldysh: New free function `conv()`
It computes a contour convolution and a sum over its corresponding target
indices of two contour function 'a' and 'b' w.r.t. one or more pairs of
arguments.
If the non-time components of the meshes of 'a' and 'b' agree, then it
uses the same non-time mesh for the result. Otherwise the result is
defined on a direct product of the non-time components of the meshes.

`KeldyshGF.__matmul__()` has been restored and now simply calls
`conv()`.